add notifications for comments, posts and reports - working, cleanup needed
