+++ citizen_report/notifications/models.py
@@ -1,8 +1,4 @@
 from django.db import models
-from django.db.models.signals import post_save
-from django.dispatch import receiver
-from django.utils import timezone
-from django.core.mail import send_mail
 from django.conf import settings
 
 from report.models import Report
@@ -22,23 +18,3 @@ class Notification(models.Model):
 
     def __repr__(self):
         return f'{self.report}, {self.user}, {self.sent_at}, {self.message}, {self.is_read}'
-
-
-# UNDER CONSTRUCTION:
-@receiver(post_save, sender=Report)
-def create_notification(sender, created, instance, update_fields, **kwargs):
-    message = f'new report {instance.title} added!' if created else f'report status or fields:{update_fields} changed'
-    notification = Notification(report=instance, user=instance.user, sent_at=timezone.now,
-                                message=message, is_read=False)
-    notification.save()
-    clerk = instance.clerk
-    send_mail(
-        f'subject: {message}',
-        f'{repr(notification)}',
-        f'{clerk.email}',
-        [f'{instance.user.email}'],
-    )
-    print(notification)
-    print(instance)
-    print(created)
-    print(update_fields)
